+++ src/main/java/racingcar/controller/RacingApp.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.controller;
 
 public class RacingApp {
