+++ src/main/java/racingcar/domain/RacingCar.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public class RacingCar {
